requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
python-dotenv
langchain_community
-e .


# ctransformers==0.2.5
# sentence-transformers==2.2.2
# pinecone-client
# langchain==0.0.225
# flask
# pypdf
# python-dotenv
# -e .
